Update docker build action, move completion command
