+++ tools/wasp-cli/completion/generator.go
@@ -1,4 +1,4 @@
-package main
+package completion
 
 import (
 	"fmt"
@@ -9,7 +9,7 @@ import (
 	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
-func completionCmd(name string) *cobra.Command {
+func CompletionCmd(name string) *cobra.Command {
 	completionCmd := &cobra.Command{
 		Use:   "gen_completion [bash|zsh|fish|powershell]",
 		Short: "Generates shell autocompletion script. Run `wasp-cli gen_completion -h` for instructions.",
